fixed configuration and initialization
