--- src/main/java/com/crsms/config/Initializer.java
@@ -1,30 +1,0 @@
-package com.crsms.config;
-
-import org.springframework.web.WebApplicationInitializer;
-import org.springframework.web.context.ContextLoaderListener;
-import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
-import org.springframework.web.servlet.DispatcherServlet;
-
-import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRegistration;
-
-//public class Initializer implements WebApplicationInitializer {
-//						
-//    private static final String DISPATCHER_SERVLET_NAME = "dispatcher";
-//
-//    @Override
-//    public void onStartup(ServletContext servletContext) throws ServletException {
-//        AnnotationConfigWebApplicationContext context = new AnnotationConfigWebApplicationContext();
-//
-//        //Register in context the config file.
-//        context.register(AppConfig.class);
-//        servletContext.addListener(new ContextLoaderListener(context));
-//        context.setServletContext(servletContext);
-//
-//        ServletRegistration.Dynamic servletRegistration = 
-//        		servletContext.addServlet(DISPATCHER_SERVLET_NAME, new DispatcherServlet(context));
-//        servletRegistration.addMapping("/");
-//        servletRegistration.setLoadOnStartup(1);
-//    }
-//}
