+++ src/main/java/com/crsms/config/RootConfig.java
@@ -1,0 +1,21 @@
+package com.crsms.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.web.servlet.ViewResolver;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+import org.springframework.web.servlet.view.InternalResourceViewResolver;
+import org.springframework.web.servlet.view.JstlView;
+
+@Configuration
+//@EnableWebMvc is turned off for console debug only!!!
+@EnableWebMvc
+@ComponentScan(basePackages = "com.crsms")
+@Import({ SecurityConfig.class, HibernateConfig.class})
+public class RootConfig extends WebMvcConfigurerAdapter {
+
+}
