+++ src/main/java/com/crsms/config/SecurityInitializer.java
@@ -2,6 +2,6 @@ package com.crsms.config;
 
 import org.springframework.security.web.context.AbstractSecurityWebApplicationInitializer;
 
-public class SpringSecurityInitializer extends AbstractSecurityWebApplicationInitializer {
+public class SecurityInitializer extends AbstractSecurityWebApplicationInitializer {
 
 }
