--- src/main/java/com/crsms/config/SpringMain.java
@@ -1,16 +1,0 @@
-package com.crsms.config;
-
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-
-public class SpringMain {
-	
-	public static void main(String[] args) {
-
-		ApplicationContext context = new AnnotationConfigApplicationContext(AppConfig.class);
-
-		// created for testing purposes and experiments with console only
-	
-		
-	}
-}
