+++ src/main/java/com/crsms/config/WebInitializer.java
@@ -2,18 +2,18 @@ package com.crsms.config;
 
 import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
 
-import com.crsms.config.AppConfig;
+import com.crsms.config.RootConfig;
 
-public class SpringMvcInitializer extends AbstractAnnotationConfigDispatcherServletInitializer {
+public class WebInitializer extends AbstractAnnotationConfigDispatcherServletInitializer {
 
 	@Override
 	protected Class<?>[] getRootConfigClasses() {
-		return new Class[] { AppConfig.class };
+		return new Class[] { RootConfig.class };
 	}
 
 	@Override
 	protected Class<?>[] getServletConfigClasses() {
-		return null;
+		return new Class[] { WebConfig.class };
 	}
 
 	@Override
